+++ tencentcloud/captcha/v20190722/models.go
@@ -654,6 +654,7 @@ type DescribeCaptchaMiniResultResponse struct {
 	// 15     ticket decryption failed     票据解密失败
 	// 16     CaptchaAppId wrong format     检查验证码应用APPID错误
 	// 21     ticket error     票据验证错误
+	// 25     invalid ticket     无效票据
 	// 26     system internal error     系统内部错误
 	// 100   param err     参数校验错误
 		CaptchaCode *int64 `json:"CaptchaCode,omitempty" name:"CaptchaCode"`