+++ tencentcloud/eis/v20210601/models.go
@@ -371,7 +371,7 @@ type RuntimeDeployedInstanceMC struct {
 	// 应用实例部署更新时间
 	UpdatedAt *int64 `json:"UpdatedAt,omitempty" name:"UpdatedAt"`
 
-	// 应用类型：0:NormalApp普通应用 1:TemplateApp模版应用 2:LightApp轻应用 3:MicroConnTemplate微连接模板 4:MicroConnApp微连接应用
+	// 应用类型：0:NormalApp普通应用 1:TemplateApp模板应用 2:LightApp轻应用 3:MicroConnTemplate微连接模板 4:MicroConnApp微连接应用
 	ProjectType *int64 `json:"ProjectType,omitempty" name:"ProjectType"`
 }